Windows port: experimental deepseek_prefill: fix narrowing conversions
- attn_res_gather_softmax: static_cast CoreCoord x/y (size_t) to
  uint32_t in SetRuntimeArgs vector
- combine, dispatch: static_cast packet_header_size_bytes (size_t) to
  uint32_t in CBFormatDescriptor::page_size designated initializer;
  static_cast MeshCoordinateRange/MeshShape accessors (size_t) to
  uint32_t in runtime-args initializer lists
- zero_padded_kv_cache: static_cast CB index (uint32_t) to uint8_t in
  CBFormatDescriptor::buffer_index designated initializer

Note: attn_res_gather_softmax, attn_res_weighted_reduce_nc,
inbound_socket_service_sync, outbound_socket_service_sync,
pack_scaled_fp8_kv_cache, per_token_cast_to_fp8,
rotary_embedding_indexed, update_padded_kv_cache, and
zero_padded_kv_cache still fail to build in this tree: their
generated .obj/.lib output paths exceed the 260-char MAX_PATH limit
under .build/windows-msvc, causing C1083/LNK1104 regardless of source
correctness. This needs a build-tree-location fix, not a source fix.

+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/dispatch/device/dispatch_program_factory.cpp
@@ -580,7 +580,7 @@ tt::tt_metal::ProgramDescriptor create_dispatch_program(
             .format_descriptors = {{tt::tt_metal::CBFormatDescriptor{
                 .buffer_index = static_cast<uint8_t>(tt::CBIndex::c_8),
                 .data_format = tt::DataFormat::UInt8,
-                .page_size = packet_header_size_bytes,
+                .page_size = static_cast<uint32_t>(packet_header_size_bytes),
             }}},
         });
     }
@@ -633,7 +633,7 @@ tt::tt_metal::ProgramDescriptor create_dispatch_program(
         detail::get_page_size(dispatch_table_tensor),
 
         // Operation parameters (7)
-        mesh_view.num_devices(),  // num_devices
+        static_cast<uint32_t>(mesh_view.num_devices()),  // num_devices
         (uint32_t)hidden_size,
         operation_attributes.experts_per_chip,
         operation_attributes.num_routed_experts,
@@ -644,8 +644,8 @@ tt::tt_metal::ProgramDescriptor create_dispatch_program(
         // Mesh information (5)
         src_mesh_id,
         src_chip_id,
-        mesh_view.num_rows(),
-        mesh_view.num_cols(),
+        static_cast<uint32_t>(mesh_view.num_rows()),
+        static_cast<uint32_t>(mesh_view.num_cols()),
         linearized_mesh_coord,
 
         // Aligned page sizes (6)
@@ -809,9 +809,9 @@ tt::tt_metal::ProgramDescriptor create_dispatch_program(
             operation_attributes.max_dispatch_buffer_token_size,   // 19
             s,                                                     // 20: dispatch_core_idx
             num_cores,                                             // 21: num_dispatch_cores
-            mesh_view.num_devices(),                               // 22: num_devices
-            mesh_view.num_rows(),                                  // 23
-            mesh_view.num_cols(),                                  // 24
+            static_cast<uint32_t>(mesh_view.num_devices()),        // 22: num_devices
+            static_cast<uint32_t>(mesh_view.num_rows()),           // 23
+            static_cast<uint32_t>(mesh_view.num_cols()),           // 24
             linearized_mesh_coord,                                 // 25
             static_cast<uint32_t>(topology),                       // 26
             block_ct_dim_dispatch,                                 // 27: must match the compute kernel
